+++ packages/yourloops/lib/notifications/models/enums/notification-type.enum.ts
@@ -26,7 +26,6 @@
  */
 
 export enum NotificationType {
-  careTeamMonitoringInvitation,
   careTeamPatientInvitation,
   careTeamProInvitation,
   directInvitation,